pyproject.toml
[tool.poetry]
name = "loaddocuments"
version = "0.1.0"
description = ""
authors = ["Nurlan-Grachyov <nurlan.grachyov@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.1"
python-dotenv = "^1.1.0"
psycopg2 = "^2.9.10"
djangorestframework = "^3.16.0"
djangorestframework-simplejwt = "^5.5.0"
drf-yasg = "^1.21.10"
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "7.1.2"
celery = "^5.5.2"
redis = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
